requirements.txt
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo==4.6.1
python-dotenv==1.0.0
